requirements.txt
flask
google-cloud-aiplatform
pyyaml
python-dotenv
